# perf: avoid double checking `value0` Future.