@He-Pin 2025-01-03
refs: https://github.com/scala/scala3/issues/22304