@He-Pin 2025-09-16
new:

```scala
[info] Benchmark                        (pool)  (recursion)  (threads)   Mode  Cnt     Score    Error   Units
[info] CompleteFutureBenchmark.failure     fjp         1024          1  thrpt    5  2480.031 ± 13.381  ops/ms
[info] CompleteFutureBenchmark.failure     fix         1024          1  thrpt    5  2463.211 ± 14.210  ops/ms
[info] CompleteFutureBenchmark.failure     fie         1024          1  thrpt    5  2479.808 ± 20.546  ops/ms
[info] CompleteFutureBenchmark.failure     gbl         1024          1  thrpt    5  2479.991 ± 26.142  ops/ms
[info] CompleteFutureBenchmark.success     fjp         1024          1  thrpt    5  2485.869 ± 12.003  ops/ms
[info] CompleteFutureBenchmark.success     fix         1024          1  thrpt    5  2485.317 ± 15.475  ops/ms
[info] CompleteFutureBenchmark.success     fie         1024          1  thrpt    5  2482.062 ±  8.531  ops/ms
[info] CompleteFutureBenchmark.success     gbl         1024          1  thrpt    5  2480.775 ±  6.382  ops/ms
```

old:

```scala
WIP
```