@He-Pin 2025-09-16
I think it is better to just keep it in a library, as you said, this changed the semantics