@lrytz 2025-09-17
@He-Pin we can discuss a change of semantics of course, but it should not be done in this PR which is only supposed to improve performance.

I moved the `case f: Future[T]` match into the extractor to avoid the `asInstanceOf[T]` on Scala 3 (see also https://github.com/scala/scala3/issues/22304#issuecomment-3302042707).